backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
requests==2.32.3
python-dotenv==1.0.1
exa_py==1.0.6
openai==1.12.0
pydantic==2.10.6
httpx==0.28.1
fastapi==0.103.2